functions/agent/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
google-api-python-client==2.*
google-auth==2.*
requests==2.*
